requirements.txt
pandas==2.3.2
streamlit==1.49.1
pillow
plotly==6.3.0
folium==0.20.0
streamlit-folium==0.25.1